added deletion conformation, sweetAlert

+++ server/public/scripts/client.js
@@ -1,41 +1,41 @@
-console.log('hello from client'); 
+console.log('hello from client');
 
-$(document).ready(readyNow); 
+$(document).ready(readyNow);
 
-function readyNow() { 
-    console.log('hello from jquery'); 
+function readyNow() {
+    console.log('hello from jquery');
     // Click handlers for add buttons 
-    $('.add-btn').on('click', addTask); 
+    $('.add-btn').on('click', addTask);
     $('#task-container').on('click', '.btn-delete', deleteTask);
     $('#task-container').on('click', '.task-td', updateStatus);
     // update tasks upon refresh 
-    getTasks(); 
-} 
-function getTasks() { 
-    $.ajax({
-        type: 'GET', 
-        url: '/list', 
-    })
-    .then(function (response){
-        console.log(response);
-        renderTasks(response)
-    })
-    .catch( function(error){
-        console.log('Error:', error);
-        alert('Something bad happened. Try again later');
-    })
+    getTasks();
 }
-function addTask() { 
+function getTasks() {
+    $.ajax({
+        type: 'GET',
+        url: '/list',
+    })
+        .then(function (response) {
+            console.log(response);
+            renderTasks(response)
+        })
+        .catch(function (error) {
+            console.log('Error:', error);
+            alert('Something bad happened. Try again later');
+        })
+}
+function addTask() {
     // get which button was clicked 
-    let listButtonClicked = $(this).attr('id'); 
+    let listButtonClicked = $(this).attr('id');
     // get the value of the input was clicked based on the button 
-    let task = $(`#${listButtonClicked}-input`).val(); 
+    let task = $(`#${listButtonClicked}-input`).val();
     // basic data validation 
     if (task == "") {
-        return 'never added input'; 
-    } 
-    console.log(`adding task to ${listButtonClicked} list...`); 
-    console.log(task); 
+        return 'never added input';
+    }
+    console.log(`adding task to ${listButtonClicked} list...`);
+    console.log(task);
     // sending the new task to the server 
     $.ajax({
         method: 'POST',
@@ -44,30 +44,48 @@ function addTask() {
             task: task,
             taskDisplay: listButtonClicked
         }
-    }).then( function(response) {
-        getTasks(); 
-        clearInputs(); 
-        console.log('getting tasks...'); 
-    }).catch( function(error){
-        console.log('Error:', error);
-        alert('Something bad happened. Try again later');
-    }) 
-}
-// delete a tasks 
-function deleteTask() { 
-    console.log(`deleting task...`); 
-    let taskId = $(this).closest('tr').data('id');  
-    $.ajax({
-        method: 'DELETE',
-        url: `/list/${taskId}` //add id to the url
-    }).then( function(response) {
+    }).then(function (response) {
         getTasks();
-    }).catch( function(error){
+        clearInputs();
+        console.log('getting tasks...');
+    }).catch(function (error) {
         console.log('Error:', error);
         alert('Something bad happened. Try again later');
     })
-} 
-function updateStatus() { 
+}
+// alert user before they delete a file
+// delete a tasks 
+function deleteTask() {
+    swal({
+        title: "Are you sure?",
+        text: "Once deleted, you will not be able to recover this task!",
+        icon: "warning",
+        buttons: true,
+        dangerMode: true,
+    })
+    .then((willDelete) => {
+        if (willDelete) {
+            swal("Poof! Your task has been deleted!", {
+                icon: "success",
+            });
+
+            console.log(`deleting task...`);
+            let taskId = $(this).closest('tr').data('id');
+            $.ajax({
+                method: 'DELETE',
+                url: `/list/${taskId}` //add id to the url
+            }).then(function (response) {
+                getTasks();
+            }).catch(function (error) {
+                console.log('Error:', error);
+                alert('Something bad happened. Try again later');
+            })
+        } else {
+            swal("Your imaginary file is safe!");
+        }
+    });
+}
+function updateStatus() {
     let taskStatus = $(this).closest('tr').data('status');
     let taskId = $(this).closest('tr').data('id');
     console.log(taskStatus)
@@ -75,52 +93,53 @@ function updateStatus() {
     $.ajax({
         method: 'PUT',
         url: `/list/${taskId}`, //add id to the url
-        data: {taskStatus: taskStatus}
-    }).then( function(response) { 
+        data: { taskStatus: taskStatus }
+    }).then(function (response) {
         getTasks();
-        console.log('getting tasks...'); 
-    }).catch( function(error){
+        console.log('getting tasks...');
+    }).catch(function (error) {
         console.log('Error:', error);
         alert('Something bad happened. Try again later');
-    }) 
-} 
+    })
+}
 
 // clear inputs 
-function clearInputs() { 
+function clearInputs() {
     $('#client-input').val('');
     $('#server-input').val('');
     $('#database-input').val('');
 }
 
 // append all of the tasks to the DOM 
-function renderTasks(allTasks) { 
+function renderTasks(allTasks) {
     // clear out data from the last render 
     $('#client-col').empty();
     $('#server-col').empty();
     $('#database-col').empty();
-    
+
     // using two for loops one for complete tasks and the other for incomplete tasks I can push completed tasks
     // to the bottom of the list. 
-    for (let task of allTasks) { 
-        if(task.taskDisplay === "client" && task.status === 'false' ) {
-            $('#client-col').append(`<tr data-id="${task.id}" data-status="${task.status}"><td class="task-td">${task.task}</td><td><button class="btn-delete">-</button></td></tr>`);
-        } 
-        else if(task.taskDisplay === "server" && task.status === 'false' ) {
-            $('#server-col').append(`<tr data-id="${task.id}" data-status="${task.status}"><td class="task-td">${task.task}</td><td><button class="btn-delete">-</button></td></tr>`);
-        } 
-        else if(task.taskDisplay === "database" && task.status === 'false' ) {
-            $('#database-col').append(`<tr data-id="${task.id}" data-status="${task.status}"><td class="task-td">${task.task}</td><td><button class="btn-delete">-</button></td></tr>`);
-        }         
-    } 
     for (let task of allTasks) {
-        if(task.taskDisplay === "client" && task.status === 'true' ) {
+        if (task.taskDisplay === "client" && task.status === 'false') {
+            $('#client-col').append(`<tr data-id="${task.id}" data-status="${task.status}"><td class="task-td">${task.task}</td><td><button class="btn-delete">-</button></td></tr>`);
+        }
+        else if (task.taskDisplay === "server" && task.status === 'false') {
+            $('#server-col').append(`<tr data-id="${task.id}" data-status="${task.status}"><td class="task-td">${task.task}</td><td><button class="btn-delete">-</button></td></tr>`);
+        }
+        else if (task.taskDisplay === "database" && task.status === 'false') {
+            $('#database-col').append(`<tr data-id="${task.id}" data-status="${task.status}"><td class="task-td">${task.task}</td><td><button class="btn-delete">-</button></td></tr>`);
+        }
+    }
+    for (let task of allTasks) {
+        if (task.taskDisplay === "client" && task.status === 'true') {
             $('#client-col').append(`<tr data-id="${task.id}" data-status="${task.status}"><td class="task-td complete">${task.task}</td><td><button class="btn-delete">-</button></td></tr>`);
-        } 
-        else if(task.taskDisplay === "server" && task.status === 'true' ) {
+        }
+        else if (task.taskDisplay === "server" && task.status === 'true') {
             $('#server-col').append(`<tr data-id="${task.id}" data-status="${task.status}"><td class="task-td complete">${task.task}</td><td><button class="btn-delete">-</button></td></tr>`);
-        } 
-        else if(task.taskDisplay === "database" && task.status === 'true' ) {
+        }
+        else if (task.taskDisplay === "database" && task.status === 'true') {
             $('#database-col').append(`<tr data-id="${task.id}" data-status="${task.status}"><td class="task-td complete">${task.task}</td><td><button class="btn-delete">-</button></td></tr>`);
-        } 
+        }
     }
 }
+
